refactor: rename button state file
